third_party/blink/web_tests/fast/js/nested-function-scope-expected.txt

This page tests issues of scope with nested functions.

Before parsing f:
-----
PASS: typeof f should be function and is.
PASS: typeof f.f1 should be undefined and is.

After parsing f, but before calling f:
-----
PASS: typeof f should be function and is.
PASS: typeof f.f1 should be undefined and is.

In call to f:
-----
PASS: typeof f should be function and is.
PASS: typeof f1 should be function and is.
PASS: typeof f.f1 should be undefined and is.
PASS: typeof x should be number and is.
PASS: typeof y should be undefined and is.
